+++ app/controllers/topics_controller.rb
@@ -3,24 +3,46 @@ class TopicsController < ApplicationController
 		@topics = Topic.all
   end
 
+	def my_bookmarks
+		@bookmarks = current_user.bookmarks
+
+		@topics = []
+
+		@bookmarks.each do |b|
+			if !@topics.include?(b.topic)
+				@topics.push(b.topic)
+			end
+		end
+
+		# Pundit Authorization
+		authorize Topic
+	end
+
   def show
 		@topic = Topic.find(params[:id])
   end
 
   def new
 		@topic = Topic.new
+		# Pundit Authorization
+		authorize @topic
   end
 
   def edit
 		@topic = Topic.find(params[:id])
+		# Pundit Authorization
+		authorize @topic
   end
 
 	def create
 		@topic = Topic.new(topic_params)
 		@topic.user = current_user
 
+		# Pundit Authorization
+		authorize @topic
+
 		if @topic.save
-			flash[:notice] = "Topic was save successfully."
+			flash[:notice] = "Topic was saved successfully."
 			redirect_to @topic
 		else
 			flash[:alert] = "There was an error saving the topic. Please try again."
@@ -30,6 +52,8 @@ class TopicsController < ApplicationController
 
 	def update
 		@topic = Topic.find(params[:id])
+		# Pundit Authorization
+		authorize @topic
 		@topic.assign_attributes(topic_params)
 
 		if @topic.save
@@ -43,6 +67,8 @@ class TopicsController < ApplicationController
 
 	def destroy
 		@topic = Topic.find(params[:id])
+		# Pundit Authorization
+		authorize @topic
 
 		if @topic.destroy
 			flash[:notice] = "The \"#{@topic.title}\" topic was deleted successfully."
